Update logic translate message by language code selected
