+++ lib/features/chat/chat.dart
@@ -1,0 +1,2 @@
+export 'ui/public_chat_screen.dart';
+export 'bloc/chat_cubit.dart';
