--- lib/features/language/cubit/language_cubit.dart
@@ -1,8 +1,0 @@
-import 'package:bloc/bloc.dart';
-import 'package:equatable/equatable.dart';
-
-part 'language_state.dart';
-
-class LanguageCubit extends Cubit<LanguageState> {
-  LanguageCubit() : super(LanguageInitial());
-}
