--- lib/features/language/ui/language_screen.dart
@@ -1,60 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:public_chat/utils/countries.dart';
-import 'package:country_flags/country_flags.dart';
-
-class LanguageScreen extends StatefulWidget {
-  const LanguageScreen({super.key});
-
-  @override
-  State<LanguageScreen> createState() => _LanguageScreenState();
-}
-
-class _LanguageScreenState extends State<LanguageScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        leading: GestureDetector(
-          onTap: () => Navigator.of(context).pop(),
-          child: const Icon(
-            Icons.arrow_back_ios_new,
-            color: Colors.black,
-            size: 24,
-          ),
-        ),
-        title: Text('Languages'),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16),
-        child: ListView.builder(
-          itemCount: countries.length,
-          itemBuilder: (_, index) => Container(
-            padding: const EdgeInsets.all(8),
-            margin: const EdgeInsets.only(bottom: 8),
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(16),
-              border: Border.all(width: 2, color: Colors.brown),
-            ),
-            child: Row(
-              children: [
-                CountryFlag.fromCountryCode(
-                  countries[index]['country_code'],
-                  shape: const Circle(),
-                  width: 32,
-                  height: 32,
-                ),
-                const SizedBox(width: 12),
-                Expanded(
-                  child: Text(
-                    countries[index]['name'],
-                    style: const TextStyle(color: Colors.black, fontSize: 16),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
