+++ lib/features/login/bloc/login_cubit.dart
@@ -8,6 +8,8 @@ import 'package:google_sign_in/google_sign_in.dart';
 import 'package:public_chat/repository/database.dart';
 import 'package:public_chat/service_locator/service_locator.dart';
 import 'package:public_chat/utils/bloc_extensions.dart';
+import 'package:public_chat/utils/constants.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 part 'login_state.dart';
 
@@ -44,14 +46,17 @@ class LoginCubit extends Cubit<LoginState> {
   }
 
   Future<void> _authenticateToFirebase(GoogleSignInAccount googleUser) async {
-    final GoogleSignInAuthentication googleAuth = await googleUser.authentication;
+    final GoogleSignInAuthentication googleAuth =
+        await googleUser.authentication;
 
-    final OAuthCredential oAuthCredential = GoogleAuthProvider.credential(accessToken: googleAuth.accessToken, idToken: googleAuth.idToken);
+    final OAuthCredential oAuthCredential = GoogleAuthProvider.credential(
+        accessToken: googleAuth.accessToken, idToken: googleAuth.idToken);
 
     final FirebaseAuth firebaseAuth = FirebaseAuth.instance;
     final Database database = ServiceLocator.instance.get<Database>();
     try {
-      final UserCredential userCredential = await firebaseAuth.signInWithCredential(oAuthCredential);
+      final UserCredential userCredential =
+          await firebaseAuth.signInWithCredential(oAuthCredential);
       final User? user = userCredential.user;
 
       if (user == null) {
@@ -73,10 +78,21 @@ class LoginCubit extends Cubit<LoginState> {
   Future<void> requestLogout() async {
     try {
       emitSafely(LogoutLoading());
-      await FirebaseAuth.instance.signOut().then((value) => emitSafely(LogoutSuccess()));
+      await FirebaseAuth.instance
+          .signOut()
+          .then((value) => emitSafely(LogoutSuccess()));
     } on PlatformException catch (e) {
       emitSafely(LoginFailed(e.toString()));
     }
+  }
+
+  bool checkCountryCodeLocalExisted() {
+    return (ServiceLocator.instance
+                .get<SharedPreferences>()
+                .get(Constants.prefCurrentCountryCode)
+                ?.toString() ??
+            '')
+        .isNotEmpty;
   }
 
   @override
