+++ lib/repository/database.dart
@@ -1,6 +1,9 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:public_chat/_shared/data/chat_data.dart';
+import 'package:public_chat/service_locator/service_locator.dart';
+import 'package:public_chat/utils/constants.dart';
+import 'package:translator/translator.dart';
 
 final class Database {
   static Database? _instance;
@@ -14,18 +17,52 @@ final class Database {
 
   final String _publicRoom = 'public';
   final String _userList = 'users';
+
+  Future<void> translateMessageByCurrentLanguageCode(
+    String languageCode,
+  ) async {
+    CollectionReference collectionRef =
+        FirebaseFirestore.instance.collection(_publicRoom);
+    QuerySnapshot querySnapshot = await collectionRef.get();
+
+    WriteBatch batch = FirebaseFirestore.instance.batch();
+
+    int batchLength = 0;
+    await Future.forEach(querySnapshot.docs, (doc) async {
+      Map<String, String> translated =
+          (doc.data() as Map<String, dynamic>)['translated'];
+      if (!translated.keys.contains(languageCode)) {
+        if (translated.keys.contains(Constants.languageCodeDefault)) {
+          final translator = ServiceLocator.instance.get<GoogleTranslator>();
+          final messageTranslated = await translator.translate(
+              translated[Constants.languageCodeDefault]!,
+              from: Constants.languageCodeDefault,
+              to: languageCode);
+          translated[languageCode] = messageTranslated.text;
+        }
+      }
+      batch.update(doc.reference, {'translated': translated});
+      batchLength++;
+      if (batchLength == Constants.lengthBatchUpdateTranslateMessage) {
+        await batch.commit();
+        batch = FirebaseFirestore.instance.batch();
+      }
+    });
+    await batch.commit();
+  }
+
   void writePublicMessage(Message message) {
     FirebaseFirestore.instance.collection(_publicRoom).add(message.toMap());
   }
 
   Query<T> getPublicChatContents<T>({
-    required FromFirestore<T> fromFirestore,
-    required ToFirestore<T> toFirestore,
+    required FromFirestore<T> fromFireStore,
+    required ToFirestore<T> toFireStore,
   }) {
     return FirebaseFirestore.instance
         .collection(_publicRoom)
         .orderBy('time', descending: true)
-        .withConverter(fromFirestore: fromFirestore, toFirestore: toFirestore);
+        .withConverter(fromFirestore: fromFireStore, toFirestore: toFireStore);
   }
 
   void saveUser(User user) {
@@ -41,7 +78,7 @@ final class Database {
         .collection(_userList)
         .doc(uid)
         .withConverter(
-            fromFirestore: _userDetailFromFirestore,
+            fromFirestore: _userDetailFromFireStore,
             toFirestore: _userDetailToFirestore)
         .get(const GetOptions(source: Source.serverAndCache));
   }
@@ -50,18 +87,19 @@ final class Database {
     return FirebaseFirestore.instance
         .collection(_userList)
         .withConverter(
-            fromFirestore: _userDetailFromFirestore,
+            fromFirestore: _userDetailFromFireStore,
             toFirestore: _userDetailToFirestore)
         .snapshots();
   }
 
   /// ###############################################################
-  /// fromFirestore and toFirestore
-  UserDetail _userDetailFromFirestore(
+  /// fromFireStore and toFireStore
+  UserDetail _userDetailFromFireStore(
     DocumentSnapshot<Map<String, dynamic>> snapshot,
     SnapshotOptions? options,
   ) =>
       UserDetail.fromMap(snapshot.id, snapshot.data() ?? {});
+
   Map<String, Object?> _userDetailToFirestore(
     UserDetail value,
     SetOptions? options,
