+++ lib/utils/constants.dart
@@ -1,2 +1,1241 @@
 class Constants {
+  static const String prefCurrentCountryCode = 'current_country_code_selected';
+
+  static const String countryCodeDefault = 'US';
+  static const String languageCodeDefault = 'en';
+
+  static const int lengthBatchUpdateTranslateMessage = 500;
+
+  static const List<Map<String, dynamic>> countries = [
+    {
+      "country_code": "AF",
+      "language_code": "ps",
+      "name": "Afghanistan",
+    },
+    {
+      "country_code": "AX",
+      "language_code": "sv",
+      "name": "Åland Islands",
+    },
+    {
+      "country_code": "AL",
+      "language_code": "sq",
+      "name": "Albania",
+    },
+    {
+      "country_code": "DZ",
+      "language_code": "ar",
+      "name": "Algeria",
+    },
+    {
+      "country_code": "AS",
+      "language_code": "en",
+      "name": "American Samoa",
+    },
+    {
+      "country_code": "AD",
+      "language_code": "ca",
+      "name": "Andorra",
+    },
+    {
+      "country_code": "AO",
+      "language_code": "ln",
+      "name": "Angola",
+    },
+    {
+      "country_code": "AI",
+      "language_code": "es",
+      "name": "Anguilla",
+    },
+    {
+      "country_code": "AG",
+      "language_code": "en",
+      "name": "Antigua and Barbuda",
+    },
+    {
+      "country_code": "AR",
+      "language_code": "es",
+      "name": "Argentina",
+    },
+    {
+      "country_code": "AM",
+      "language_code": "hy",
+      "name": "Armenia",
+    },
+    {
+      "country_code": "AW",
+      "language_code": "nl",
+      "name": "Aruba",
+    },
+    {
+      "country_code": "AC",
+      "language_code": "en",
+      "name": "Ascension Island",
+    },
+    {
+      "country_code": "AU",
+      "language_code": "en",
+      "name": "Australia",
+    },
+    {
+      "country_code": "AT",
+      "language_code": "de",
+      "name": "Austria",
+    },
+    {
+      "country_code": "AZ",
+      "language_code": "az",
+      "name": "Azerbaijan",
+    },
+    {
+      "country_code": "BS",
+      "language_code": "es",
+      "name": "Bahamas",
+    },
+    {
+      "country_code": "BH",
+      "language_code": "ar",
+      "name": "Bahrain",
+    },
+    {
+      "country_code": "BD",
+      "language_code": "bn",
+      "name": "Bangladesh",
+    },
+    {
+      "country_code": "BB",
+      "language_code": "es",
+      "name": "Barbados",
+    },
+    {
+      "country_code": "BY",
+      "language_code": "be",
+      "name": "Belarus",
+    },
+    {
+      "country_code": "BE",
+      "language_code": "nl",
+      "name": "Belgium",
+    },
+    {
+      "country_code": "BZ",
+      "language_code": "es",
+      "name": "Belize",
+    },
+    {
+      "country_code": "BJ",
+      "language_code": "fr",
+      "name": "Benin",
+    },
+    {
+      "country_code": "BM",
+      "language_code": "es",
+      "name": "Bermuda",
+    },
+    {
+      "country_code": "BT",
+      "language_code": "dz",
+      "name": "Bhutan",
+    },
+    {
+      "country_code": "BO",
+      "language_code": "qu",
+      "name": "Bolivia",
+    },
+    {
+      "country_code": "BA",
+      "language_code": "bs",
+      "name": "Bosnia and Herzegovina",
+    },
+    {
+      "country_code": "BW",
+      "language_code": "tn",
+      "name": "Botswana",
+    },
+    {
+      "country_code": "BR",
+      "language_code": "es",
+      "name": "Brazil",
+    },
+    {
+      "country_code": "IO",
+      "language_code": "es",
+      "name": "British Indian Ocean Territory",
+    },
+    {
+      "country_code": "VG",
+      "language_code": "es",
+      "name": "British Virgin Islands",
+    },
+    {
+      "country_code": "BN",
+      "language_code": "ms",
+      "name": "Brunei",
+    },
+    {
+      "country_code": "BG",
+      "language_code": "bg",
+      "name": "Bulgaria",
+    },
+    {
+      "country_code": "BF",
+      "language_code": "fr",
+      "name": "Burkina Faso",
+    },
+    {
+      "country_code": "BI",
+      "language_code": "fr",
+      "name": "Burundi",
+    },
+    {
+      "country_code": "KH",
+      "language_code": "km",
+      "name": "Cambodia",
+    },
+    {
+      "country_code": "CM",
+      "language_code": "agq",
+      "name": "Cameroon",
+    },
+    {
+      "country_code": "CA",
+      "language_code": "en",
+      "name": "Canada",
+    },
+    {
+      "country_code": "CV",
+      "language_code": "kea",
+      "name": "Cape Verde",
+    },
+    {
+      "country_code": "BQ",
+      "language_code": "nl",
+      "name": "Caribbean Netherlands",
+    },
+    {
+      "country_code": "KY",
+      "language_code": "es",
+      "name": "Cayman Islands",
+    },
+    {
+      "country_code": "CF",
+      "language_code": "fr",
+      "name": "Central African Republic",
+    },
+    {
+      "country_code": "TD",
+      "language_code": "ar",
+      "name": "Chad",
+    },
+    {
+      "country_code": "CL",
+      "language_code": "en",
+      "name": "Chile",
+    },
+    {
+      "country_code": "CN",
+      "language_code": "zh",
+      "name": "China",
+    },
+    {
+      "country_code": "CX",
+      "language_code": "en",
+      "name": "Christmas Island",
+    },
+    {
+      "country_code": "CC",
+      "language_code": "en",
+      "name": "Cocos [Keeling] Islands",
+    },
+    {
+      "country_code": "CO",
+      "language_code": "en",
+      "name": "Colombia",
+    },
+    {
+      "country_code": "KM",
+      "language_code": "ar",
+      "name": "Comoros",
+    },
+    {
+      "country_code": "CK",
+      "language_code": "en",
+      "name": "Cook Islands",
+    },
+    {
+      "country_code": "CD",
+      "language_code": "fr",
+      "name": "Democratic Republic Congo",
+    },
+    {
+      "country_code": "CG",
+      "language_code": "fr",
+      "name": "Republic of Congo",
+    },
+    {
+      "country_code": "CR",
+      "language_code": "es",
+      "name": "Costa Rica",
+    },
+    {
+      "country_code": "CI",
+      "language_code": "fr",
+      "name": "Côte d'Ivoire",
+    },
+    {
+      "country_code": "HR",
+      "language_code": "hr",
+      "name": "Croatia",
+    },
+    {
+      "country_code": "CU",
+      "language_code": "es",
+      "name": "Cuba",
+    },
+    {
+      "country_code": "CW",
+      "language_code": "nl",
+      "name": "Curaçao",
+    },
+    {
+      "country_code": "CY",
+      "language_code": "en",
+      "name": "Cyprus",
+    },
+    {
+      "country_code": "CZ",
+      "language_code": "cs",
+      "name": "Czech Republic",
+    },
+    {
+      "country_code": "DK",
+      "language_code": "en",
+      "name": "Denmark",
+    },
+    {
+      "country_code": "DJ",
+      "language_code": "ar",
+      "name": "Djibouti",
+    },
+    {
+      "country_code": "DM",
+      "language_code": "en",
+      "name": "Dominica",
+    },
+    {
+      "country_code": "DO",
+      "language_code": "es",
+      "name": "Dominican Republic",
+    },
+    {
+      "country_code": "TL",
+      "language_code": "pt",
+      "name": "East Timor",
+    },
+    {
+      "country_code": "EC",
+      "language_code": "qu",
+      "name": "Ecuador",
+    },
+    {
+      "country_code": "EG",
+      "language_code": "ar",
+      "name": "Egypt",
+    },
+    {
+      "country_code": "SV",
+      "language_code": "es",
+      "name": "El Salvador",
+    },
+    {
+      "country_code": "GQ",
+      "language_code": "fr",
+      "name": "Equatorial Guinea",
+    },
+    {
+      "country_code": "ER",
+      "language_code": "ar",
+      "name": "Eritrea",
+    },
+    {
+      "country_code": "EE",
+      "language_code": "et",
+      "name": "Estonia",
+    },
+    {
+      "country_code": "SZ",
+      "language_code": "ss",
+      "name": "Eswatini",
+    },
+    {
+      "country_code": "ET",
+      "language_code": "am",
+      "name": "Ethiopia",
+    },
+    {
+      "country_code": "FK",
+      "language_code": "es",
+      "name": "Falkland Islands [Islas Malvinas]",
+    },
+    {
+      "country_code": "FO",
+      "language_code": "fo",
+      "name": "Faroe Islands",
+    },
+    {
+      "country_code": "FJ",
+      "language_code": "en",
+      "name": "Fiji",
+    },
+    {
+      "country_code": "FI",
+      "language_code": "es",
+      "name": "Finland",
+    },
+    {
+      "country_code": "FR",
+      "language_code": "fr",
+      "name": "France",
+    },
+    {
+      "country_code": "GF",
+      "language_code": "fr",
+      "name": "French Guiana",
+    },
+    {
+      "country_code": "PF",
+      "language_code": "fr",
+      "name": "French Polynesia",
+    },
+    {
+      "country_code": "GA",
+      "language_code": "fr",
+      "name": "Gabon",
+    },
+    {
+      "country_code": "GM",
+      "language_code": "ff",
+      "name": "Gambia",
+    },
+    {
+      "country_code": "GE",
+      "language_code": "ka",
+      "name": "Georgia",
+    },
+    {
+      "country_code": "DE",
+      "language_code": "de",
+      "name": "Germany",
+    },
+    {
+      "country_code": "GH",
+      "language_code": "ak",
+      "name": "Ghana",
+    },
+    {
+      "country_code": "GI",
+      "language_code": "en",
+      "name": "Gibraltar",
+    },
+    {
+      "country_code": "GR",
+      "language_code": "el",
+      "name": "Greece",
+    },
+    {
+      "country_code": "GL",
+      "language_code": "da",
+      "name": "Greenland",
+    },
+    {
+      "country_code": "GD",
+      "language_code": "es",
+      "name": "Grenada",
+    },
+    {
+      "country_code": "GP",
+      "language_code": "fr",
+      "name": "Guadeloupe",
+    },
+    {
+      "country_code": "GU",
+      "language_code": "en",
+      "name": "Guam",
+    },
+    {
+      "country_code": "GT",
+      "language_code": "es",
+      "name": "Guatemala",
+    },
+    {
+      "country_code": "GG",
+      "language_code": "ee",
+      "name": "Guernsey",
+    },
+    {
+      "country_code": "GN",
+      "language_code": "fr",
+      "name": "Guinea Conakry",
+    },
+    {
+      "country_code": "GW",
+      "language_code": "ff",
+      "name": "Guinea-Bissau",
+    },
+    {
+      "country_code": "GY",
+      "language_code": "es",
+      "name": "Guyana",
+    },
+    {
+      "country_code": "HT",
+      "language_code": "fr",
+      "name": "Haiti",
+    },
+    {
+      "country_code": "HM",
+      "language_code": "en",
+      "name": "Heard Island and McDonald Islands",
+    },
+    {
+      "country_code": "HN",
+      "language_code": "es",
+      "name": "Honduras",
+    },
+    {
+      "country_code": "HK",
+      "language_code": "zh",
+      "name": "Hong Kong",
+    },
+    {
+      "country_code": "HU",
+      "language_code": "hu",
+      "name": "Hungary",
+    },
+    {
+      "country_code": "IS",
+      "language_code": "is",
+      "name": "Iceland",
+    },
+    {
+      "country_code": "IN",
+      "language_code": "as",
+      "name": "India",
+    },
+    {
+      "country_code": "ID",
+      "language_code": "id",
+      "name": "Indonesia",
+    },
+    {
+      "country_code": "IR",
+      "language_code": "ckb",
+      "name": "Iran",
+    },
+    {
+      "country_code": "IQ",
+      "language_code": "ar",
+      "name": "Iraq",
+    },
+    {
+      "country_code": "IE",
+      "language_code": "ga",
+      "name": "Ireland",
+    },
+    {
+      "country_code": "IM",
+      "language_code": "gv",
+      "name": "Isle of Man",
+    },
+    {
+      "country_code": "IL",
+      "language_code": "ar",
+      "name": "Israel",
+    },
+    {
+      "country_code": "IT",
+      "language_code": "ca",
+      "name": "Italy",
+    },
+    {
+      "country_code": "JM",
+      "language_code": "en",
+      "name": "Jamaica",
+    },
+    {
+      "country_code": "JP",
+      "language_code": "ja",
+      "name": "Japan",
+    },
+    {
+      "country_code": "JE",
+      "language_code": "en",
+      "name": "Jersey",
+    },
+    {
+      "country_code": "JO",
+      "language_code": "ar",
+      "name": "Jordan",
+    },
+    {
+      "country_code": "KZ",
+      "language_code": "kk",
+      "name": "Kazakhstan",
+    },
+    {
+      "country_code": "KE",
+      "language_code": "ebu",
+      "name": "Kenya",
+    },
+    {
+      "country_code": "KI",
+      "language_code": "en",
+      "name": "Kiribati",
+    },
+    {
+      "country_code": "XK",
+      "language_code": "sq",
+      "name": "Kosovo",
+    },
+    {
+      "country_code": "KW",
+      "language_code": "ar",
+      "name": "Kuwait",
+    },
+    {
+      "country_code": "KG",
+      "language_code": "ky",
+      "name": "Kyrgyzstan",
+    },
+    {
+      "country_code": "LA",
+      "language_code": "lo",
+      "name": "Laos",
+    },
+    {
+      "country_code": "LV",
+      "language_code": "lv",
+      "name": "Latvia",
+    },
+    {
+      "country_code": "LB",
+      "language_code": "ar",
+      "name": "Lebanon",
+    },
+    {
+      "country_code": "LS",
+      "language_code": "st",
+      "name": "Lesotho",
+    },
+    {
+      "country_code": "LR",
+      "language_code": "ff",
+      "name": "Liberia",
+    },
+    {
+      "country_code": "LY",
+      "language_code": "ar",
+      "name": "Libya",
+    },
+    {
+      "country_code": "LI",
+      "language_code": "de",
+      "name": "Liechtenstein",
+    },
+    {
+      "country_code": "LT",
+      "language_code": "lt",
+      "name": "Lithuania",
+    },
+    {
+      "country_code": "LU",
+      "language_code": "lb",
+      "name": "Luxembourg",
+    },
+    {
+      "country_code": "MO",
+      "language_code": "zh",
+      "name": "Macau",
+    },
+    {
+      "country_code": "MK",
+      "language_code": "mk",
+      "name": "Macedonia",
+    },
+    {
+      "country_code": "MG",
+      "language_code": "mg",
+      "name": "Madagascar",
+    },
+    {
+      "country_code": "MW",
+      "language_code": "ny",
+      "name": "Malawi",
+    },
+    {
+      "country_code": "MY",
+      "language_code": "ms",
+      "name": "Malaysia",
+    },
+    {
+      "country_code": "MV",
+      "language_code": "dv",
+      "name": "Maldives",
+    },
+    {
+      "country_code": "ML",
+      "language_code": "bm",
+      "name": "Mali",
+    },
+    {
+      "country_code": "MT",
+      "language_code": "mt",
+      "name": "Malta",
+    },
+    {
+      "country_code": "MH",
+      "language_code": "en",
+      "name": "Marshall Islands",
+    },
+    {
+      "country_code": "MQ",
+      "language_code": "fr",
+      "name": "Martinique",
+    },
+    {
+      "country_code": "MR",
+      "language_code": "ar",
+      "name": "Mauritania",
+    },
+    {
+      "country_code": "MU",
+      "language_code": "mfe",
+      "name": "Mauritius",
+    },
+    {
+      "country_code": "YT",
+      "language_code": "fr",
+      "name": "Mayotte",
+    },
+    {
+      "country_code": "MX",
+      "language_code": "en",
+      "name": "Mexico",
+    },
+    {
+      "country_code": "FM",
+      "language_code": "en",
+      "name": "Micronesia",
+    },
+    {
+      "country_code": "MD",
+      "language_code": "ro",
+      "name": "Moldova",
+    },
+    {
+      "country_code": "MC",
+      "language_code": "fr",
+      "name": "Monaco",
+    },
+    {
+      "country_code": "MN",
+      "language_code": "mn",
+      "name": "Mongolia",
+    },
+    {
+      "country_code": "ME",
+      "language_code": "sr",
+      "name": "Montenegro",
+    },
+    {
+      "country_code": "MS",
+      "language_code": "es",
+      "name": "Montserrat",
+    },
+    {
+      "country_code": "MA",
+      "language_code": "zgh",
+      "name": "Morocco",
+    },
+    {
+      "country_code": "MZ",
+      "language_code": "mgh",
+      "name": "Mozambique",
+    },
+    {
+      "country_code": "MM",
+      "language_code": "my",
+      "name": "Myanmar [Burma]",
+    },
+    {
+      "country_code": "NA",
+      "language_code": "af",
+      "name": "Namibia",
+    },
+    {
+      "country_code": "NR",
+      "language_code": "en",
+      "name": "Nauru",
+    },
+    {
+      "country_code": "NP",
+      "language_code": "ne",
+      "name": "Nepal",
+    },
+    {
+      "country_code": "NL",
+      "language_code": "nl",
+      "name": "Netherlands",
+    },
+    {
+      "country_code": "NC",
+      "language_code": "fr",
+      "name": "New Caledonia",
+    },
+    {
+      "country_code": "NZ",
+      "language_code": "en",
+      "name": "New Zealand",
+    },
+    {
+      "country_code": "NI",
+      "language_code": "es",
+      "name": "Nicaragua",
+    },
+    {
+      "country_code": "NE",
+      "language_code": "fr",
+      "name": "Niger",
+    },
+    {
+      "country_code": "NG",
+      "language_code": "ff",
+      "name": "Nigeria",
+    },
+    {
+      "country_code": "NU",
+      "language_code": "en",
+      "name": "Niue",
+    },
+    {
+      "country_code": "NF",
+      "language_code": "en",
+      "name": "Norfolk Island",
+    },
+    {
+      "country_code": "KP",
+      "language_code": "ko",
+      "name": "North Korea",
+    },
+    {
+      "country_code": "MP",
+      "language_code": "en",
+      "name": "Northern Mariana Islands",
+    },
+    {
+      "country_code": "NO",
+      "language_code": "se",
+      "name": "Norway",
+    },
+    {
+      "country_code": "OM",
+      "language_code": "ar",
+      "name": "Oman",
+    },
+    {
+      "country_code": "PK",
+      "language_code": "ps",
+      "name": "Pakistan",
+    },
+    {
+      "country_code": "PW",
+      "language_code": "en",
+      "name": "Palau",
+    },
+    {
+      "country_code": "PS",
+      "language_code": "ar",
+      "name": "Palestinian Territories",
+    },
+    {
+      "country_code": "PA",
+      "language_code": "es",
+      "name": "Panama",
+    },
+    {
+      "country_code": "PG",
+      "language_code": "en",
+      "name": "Papua New Guinea",
+    },
+    {
+      "country_code": "PY",
+      "language_code": "gn",
+      "name": "Paraguay",
+    },
+    {
+      "country_code": "PE",
+      "language_code": "qu",
+      "name": "Peru",
+    },
+    {
+      "country_code": "PH",
+      "language_code": "ceb",
+      "name": "Philippines",
+    },
+    {
+      "country_code": "PL",
+      "language_code": "pl",
+      "name": "Poland",
+    },
+    {
+      "country_code": "PT",
+      "language_code": "pt",
+      "name": "Portugal",
+    },
+    {
+      "country_code": "PR",
+      "language_code": "es",
+      "name": "Puerto Rico",
+    },
+    {
+      "country_code": "QA",
+      "language_code": "ar",
+      "name": "Qatar",
+    },
+    {
+      "country_code": "RE",
+      "language_code": "fr",
+      "name": "Réunion",
+    },
+    {
+      "country_code": "RO",
+      "language_code": "ro",
+      "name": "Romania",
+    },
+    {
+      "country_code": "RU",
+      "language_code": "ru",
+      "name": "Russia",
+    },
+    {
+      "country_code": "RW",
+      "language_code": "rw",
+      "name": "Rwanda",
+    },
+    {
+      "country_code": "BL",
+      "language_code": "fr",
+      "name": "Saint Barthélemy",
+    },
+    {
+      "country_code": "SH",
+      "language_code": "en",
+      "name": "Saint Helena",
+    },
+    {
+      "country_code": "KN",
+      "language_code": "en",
+      "name": "St. Kitts",
+    },
+    {
+      "country_code": "LC",
+      "language_code": "en",
+      "name": "St. Lucia",
+    },
+    {
+      "country_code": "MF",
+      "language_code": "en",
+      "name": "Saint Martin",
+    },
+    {
+      "country_code": "PM",
+      "language_code": "fr",
+      "name": "Saint Pierre and Miquelon",
+    },
+    {
+      "country_code": "VC",
+      "language_code": "en",
+      "name": "St. Vincent",
+    },
+    {
+      "country_code": "WS",
+      "language_code": "en",
+      "name": "Samoa",
+    },
+    {
+      "country_code": "SM",
+      "language_code": "it",
+      "name": "San Marino",
+    },
+    {
+      "country_code": "ST",
+      "language_code": "pt",
+      "name": "São Tomé and Príncipe",
+    },
+    {
+      "country_code": "SA",
+      "language_code": "ar",
+      "name": "Saudi Arabia",
+    },
+    {
+      "country_code": "SN",
+      "language_code": "ff",
+      "name": "Senegal",
+    },
+    {
+      "country_code": "RS",
+      "language_code": "sr",
+      "name": "Serbia",
+    },
+    {
+      "country_code": "SC",
+      "language_code": "fr",
+      "name": "Seychelles",
+    },
+    {
+      "country_code": "SL",
+      "language_code": "ff",
+      "name": "Sierra Leone",
+    },
+    {
+      "country_code": "SG",
+      "language_code": "zh",
+      "name": "Singapore",
+    },
+    {
+      "country_code": "SX",
+      "language_code": "nl",
+      "name": "Sint Maarten",
+    },
+    {
+      "country_code": "SK",
+      "language_code": "sk",
+      "name": "Slovakia",
+    },
+    {
+      "country_code": "SI",
+      "language_code": "sl",
+      "name": "Slovenia",
+    },
+    {
+      "country_code": "SB",
+      "language_code": "en",
+      "name": "Solomon Islands",
+    },
+    {
+      "country_code": "SO",
+      "language_code": "ar",
+      "name": "Somalia",
+    },
+    {
+      "country_code": "ZA",
+      "language_code": "af",
+      "name": "South Africa",
+    },
+    {
+      "country_code": "GS",
+      "language_code": "en",
+      "name": "South Georgia and the South Sandwich Islands",
+    },
+    {
+      "country_code": "KR",
+      "language_code": "ko",
+      "name": "South Korea",
+    },
+    {
+      "country_code": "SS",
+      "language_code": "ar",
+      "name": "South Sudan",
+    },
+    {
+      "country_code": "ES",
+      "language_code": "es",
+      "name": "Spain",
+    },
+    {
+      "country_code": "LK",
+      "language_code": "si",
+      "name": "Sri Lanka",
+    },
+    {
+      "country_code": "SD",
+      "language_code": "en",
+      "name": "Sudan",
+    },
+    {
+      "country_code": "SR",
+      "language_code": "nl",
+      "name": "Suriname",
+    },
+    {
+      "country_code": "SJ",
+      "language_code": "nb",
+      "name": "Svalbard and Jan Mayen",
+    },
+    {
+      "country_code": "SE",
+      "language_code": "sv",
+      "name": "Sweden",
+    },
+    {
+      "country_code": "CH",
+      "language_code": "fr",
+      "name": "Switzerland",
+    },
+    {
+      "country_code": "SY",
+      "language_code": "syr",
+      "name": "Syria",
+    },
+    {
+      "country_code": "TW",
+      "language_code": "zh",
+      "name": "Taiwan",
+    },
+    {
+      "country_code": "TJ",
+      "language_code": "tg",
+      "name": "Tajikistan",
+    },
+    {
+      "country_code": "TZ",
+      "language_code": "asa",
+      "name": "Tanzania",
+    },
+    {
+      "country_code": "TH",
+      "language_code": "th",
+      "name": "Thailand",
+    },
+    {
+      "country_code": "TG",
+      "language_code": "ee",
+      "name": "Togo",
+    },
+    {
+      "country_code": "TK",
+      "language_code": "en",
+      "name": "Tokelau",
+    },
+    {
+      "country_code": "TO",
+      "language_code": "to",
+      "name": "Tonga",
+    },
+    {
+      "country_code": "TT",
+      "language_code": "es",
+      "name": "Trinidad/Tobago",
+    },
+    {
+      "country_code": "TN",
+      "language_code": "ar",
+      "name": "Tunisia",
+    },
+    {
+      "country_code": "TR",
+      "language_code": "tr",
+      "name": "Turkey",
+    },
+    {
+      "country_code": "TM",
+      "language_code": "tk",
+      "name": "Turkmenistan",
+    },
+    {
+      "country_code": "TC",
+      "language_code": "es",
+      "name": "Turks and Caicos Islands",
+    },
+    {
+      "country_code": "TV",
+      "language_code": "en",
+      "name": "Tuvalu",
+    },
+    {
+      "country_code": "VI",
+      "language_code": "es",
+      "name": "U.S. Virgin Islands",
+    },
+    {
+      "country_code": "UG",
+      "language_code": "lg",
+      "name": "Uganda",
+    },
+    {
+      "country_code": "UA",
+      "language_code": "uk",
+      "name": "Ukraine",
+    },
+    {
+      "country_code": "AE",
+      "language_code": "ar",
+      "name": "United Arab Emirates",
+    },
+    {
+      "country_code": "GB",
+      "language_code": "en",
+      "name": "United Kingdom",
+    },
+    {
+      "country_code": "US",
+      "language_code": "en",
+      "name": "United States",
+    },
+    {
+      "country_code": "UY",
+      "language_code": "es",
+      "name": "Uruguay",
+    },
+    {
+      "country_code": "UZ",
+      "language_code": "uz",
+      "name": "Uzbekistan",
+    },
+    {
+      "country_code": "VU",
+      "language_code": "fr",
+      "name": "Vanuatu",
+    },
+    {
+      "country_code": "VA",
+      "language_code": "it",
+      "name": "Vatican City",
+    },
+    {
+      "country_code": "VE",
+      "language_code": "es",
+      "name": "Venezuela",
+    },
+    {
+      "country_code": "VN",
+      "language_code": "vi",
+      "name": "Vietnam",
+    },
+    {
+      "country_code": "WF",
+      "language_code": "fr",
+      "name": "Wallis and Futuna",
+    },
+    {
+      "country_code": "EH",
+      "language_code": "ar",
+      "name": "Western Sahara",
+    },
+    {
+      "country_code": "YE",
+      "language_code": "ar",
+      "name": "Yemen",
+    },
+    {
+      "country_code": "ZM",
+      "language_code": "bem",
+      "name": "Zambia",
+    },
+    {
+      "country_code": "ZW",
+      "language_code": "nd",
+      "name": "Zimbabwe",
+    }
+  ];
 }
