+++ pubspec.yaml
@@ -33,7 +33,6 @@ dependencies:
   shared_preferences: ^2.2.3
 
   country_flags: ^3.0.0
-  country_picker: ^2.0.26
   translator: ^1.0.0
 
 dev_dependencies:
